requirements.txt
pytest
pytest-playwright
pytest-bdd
python-dotenv
allure-pytest

